requirements.txt
# Core dependencies (using compatible versions for Python 3.13)
pyspark>=3.3.0
kafka-python>=2.0.2
tweepy>=4.14.0
pymysql>=1.1.0
mysql-connector-python>=8.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Natural Language Processing
nltk>=3.8.0
textblob>=0.17.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0

# Additional
# Note: pyspark-streaming-kafka may need to be installed separately or via Spark packages

# PHP Twitter API (install via Composer)
# composer require abraham/twitteroauth

